disasm: Fix referencing the incorrect function type in call_indirect
In Disassemble, the function uses the integer immediate for
call_indirect as an index into the global index space, while it should
be an index into the Type Section of the module. This causes  it to
reference an incorrect function, which would cause stack underflows
later on. This CL fixes that.

Also, fix the select operator decreasing the stack height by 3 instead
of 2.

(Fixes #49)